[FE/HOTFIX] 연결 시 발생한 에러를 해결합니다.

+++ FE/src/apis/instance.ts
@@ -8,13 +8,13 @@ import {
 import ERROR_CODE from "@/constants/ERROR_CODE";
 import ERROR_MESSAGE from "@/constants/ERROR_MESSAGE";
 import { toast } from "react-toastify";
-import { redirect } from "next/navigation";
 
 const https = axios.create({
   baseURL: process.env.NEXT_PUBLIC_API_URL + "/api",
   headers: {
     "Content-Type": "application/json",
   },
+  withCredentials: true,
 });
 
 https.interceptors.request.use(
@@ -63,12 +63,21 @@ https.interceptors.response.use(
       return await axios(originalRequest);
     }
 
+    if (errorCode === ERROR_CODE.AUTH.INVALID_NAME) {
+      toast.error(errorMessage, {
+        toastId: errorCode,
+      });
+      return Promise.reject(error);
+    }
+
     if (Object.values(ERROR_CODE.AUTH).includes(errorCode)) {
-      toast.error(errorMessage);
+      toast.error(errorMessage, {
+        toastId: errorCode,
+      });
       deleteTokenInfo();
       setTimeout(() => {
         window.location.href = "/login";
-      }, 3000); // 2초 후에 페이지 이동
+      }, 3000);
     }
     error.message = errorMessage;
     return Promise.reject(error);
